backend/pyproject.toml
[project]
name = "quant-dashboard-backend"
version = "0.1.0"
description = "FastAPI backend for quantitative research dashboard"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.36",
  "asyncpg>=0.30.0",
  "redis>=5.2.0",
  "httpx>=0.28.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "loguru>=0.7.2",
  "yfinance>=0.2.54",
  "pandas>=2.2.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.4"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
